📁Add: apply FSD pattern to improve folder structure
