+++ src/entities/expo-detail/index.tsx
@@ -1,0 +1,5 @@
+export { default as ContentText } from './ui/ContentText';
+export { default as DetailHeader } from './ui/DetailHeader';
+export { default as ExpoActionPanel } from './ui/ExpoActionPanel';
+export { default as KaKaoMap } from './ui/KaKaoMap';
+export { default as QRcode } from './ui/QRcode';
