+++ src/entities/expo-detail/ui/DetailHeader/index.tsx
@@ -4,7 +4,11 @@ import { useRouter } from 'next/navigation';
 import React from 'react';
 import { ArrowLeft } from '@/shared/assets/icons';
 
-const DetailHeader = ({ HederTitle }: { HederTitle: string }) => {
+interface Props {
+  headerTitle: string;
+}
+
+const DetailHeader = ({ headerTitle }: Props) => {
   const router = useRouter();
 
   return (
@@ -12,7 +16,7 @@ const DetailHeader = ({ HederTitle }: { HederTitle: string }) => {
       <label className="hover:cursor-pointer" onClick={() => router.back()}>
         <ArrowLeft />
       </label>
-      <p className="text-h1 text-black mobile:text-body1">{HederTitle}</p>
+      <p className="text-h1 text-black mobile:text-body1">{headerTitle}</p>
     </div>
   );
 };
