+++ src/entities/expo-detail/ui/KaKaoMap/index.tsx
@@ -1,38 +1,15 @@
 'use client';
 
 import Script from 'next/script';
-import { useEffect, useRef } from 'react';
+import { useKakaoMap } from '../../model/useKakaoMap';
 
 interface KakaoMapProps {
   latitude: number;
   longitude: number;
 }
 
-export default function KakaoMap({ latitude, longitude }: KakaoMapProps) {
-  const mapRef = useRef<HTMLDivElement>(null);
-
-  const loadKakaoMap = () => {
-    if (typeof window.kakao === 'undefined' || !mapRef.current) return;
-
-    const position = new window.kakao.maps.LatLng(latitude, longitude);
-
-    const options = {
-      center: position,
-      level: 3,
-    };
-
-    const map = new window.kakao.maps.Map(mapRef.current, options);
-
-    const marker = new window.kakao.maps.Marker({
-      position: position,
-    });
-
-    marker.setMap(map);
-  };
-
-  useEffect(() => {
-    loadKakaoMap();
-  }, [latitude, longitude]);
+const KakaoMap = ({ latitude, longitude }: KakaoMapProps) => {
+  const { mapRef, loadKakaoMap } = useKakaoMap({ latitude, longitude });
 
   return (
     <>
@@ -44,4 +21,6 @@ export default function KakaoMap({ latitude, longitude }: KakaoMapProps) {
       <div ref={mapRef} className="h-[400px] w-full" />
     </>
   );
-}
+};
+
+export default KakaoMap;
