+++ src/entities/expo-detail/ui/QRcode/index.stories.ts
