+++ src/entities/expo-detail/ui/QRcode/index.tsx
