+++ src/shared/ui/index.tsx
@@ -1,3 +1,4 @@
 export { default as Button } from './Button';
 export { default as Input } from './Input';
 export { default as Modal } from './Modal';
+export { default as WhiteButton } from './WhiteButton';
