+++ src/widgets/expo-detail/index.tsx
@@ -1,0 +1,1 @@
+export { default as ExpoDetailLayout } from './ui/ExpoDetailLayout';
